backend/rest-api/pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = ""
authors = ["nikosrk <nrekkas@gmail.com>"]
# readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
datetime = "^5.5"
sqlmodel = "^0.0.16"
pytz = "^2024.1"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.1"
pyjwt = "^2.8.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-multipart = "^0.0.9"
fastapi-cors = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
